api.game/requirements.txt
flask
flask-restful
flask-mongoengine
pyyaml
requests
pymongo